pyproject.toml
[project]
name = "grafana-dashboard-builder"
version = "0.10.0a3"
description = "Generate Grafana dashboards with YAML"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "argparse",
    "requests",
    "requests-kerberos",
    "PyYAML>=5.3",
    "typing-extensions>=4.15.0",
]
keywords = ["grafana yaml graphite prometheus influxdb"]
authors = [
    {name = "Jakub Plichta", email = "jakub.plichta@gmail.com"},
]
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Utilities",
]
license-files = [
    "LICENSE",
    "NOTICE",
    "AUTHORS",
]

[project.urls]
Homepage = "https://github.com/jakubplichta/grafana-dashboard-builder"

[project.scripts]
grafana-dashboard-builder = "grafana_dashboards.cli:main"
"grafana_dashboard_builder.py" = "grafana_dashboards.cli:main"

[dependency-groups]
dev = [
    "mock>=5.2.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.0",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
]

[build-system]
requires = ["setuptools>=78.1.0", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "grafana_dashboards",
    "grafana_dashboards.client",
    "grafana_dashboards.components"
]

[tool.mypy]
exclude = "tests"
# Start off with these
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true

# Getting this passing should be easy
strict_equality = true

# Strongly recommend enabling this one as soon as you can
check_untyped_defs = true

# These shouldn't be too much additional work, but may be tricky to
# get passing if you use a lot of untyped libraries
disallow_subclassing_any = true
disallow_untyped_decorators = true
disallow_any_generics = true

# These next few are various gradations of forcing use of type annotations
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_defs = true

# This one isn't too hard to get passing, but return on investment is lower
no_implicit_reexport = true

# This one can be tricky to get passing if you use a lot of untyped libraries
#warn_return_any = true

# This one is a catch-all flag for the rest of strict checks that are technically
# correct but may not be practical
extra_checks = true

[[tool.mypy.overrides]]
module = ["requests_kerberos.*", "grafana_dashboards.gdbyaml"]
ignore_missing_imports = true
ignore_errors = true

[tool.tox]
requires = ["tox>=4.31.0"]
env_list = ["ruff", "mypy", "py3.9", "py3.10", "py3.11", "py3.12", "py3.13", "py3.14"]

[tool.tox.env_run_base]
description = "Run test under {base_python}"
runner = "uv-venv-lock-runner"
dependency_groups = ["dev"]
commands = [["pytest", "--cov=grafana_dashboards"]]

[tool.tox.env.ruff]
description = "Run code style checks"
commands = [["ruff", "check"]]

[tool.tox.env.mypy]
description = "Run code type checks"
commands = [["mypy", "grafana_dashboards"]]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 160

[tool.ruff.lint]
#select = ["E", "F", "UP", "B", "SIM", "I"]
select = ["E", "F", "UP", "B"]
ignore = []
preview = true
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
indent-style = "space"
quote-style = "single"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
